requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0

# SAM — Segment Anything Model (Meta)
git+https://github.com/facebookresearch/segment-anything.git

# Recommended (significant augmentation quality improvement)
albumentations>=1.3.0

# COCO RLE mask decoding (optional but recommended)
pycocotools>=2.0.7
